pyproject.toml
[tool.poetry]
name = "django-quiz"
version = "0.1.0"
description = "Sample application to build a quiz with Django"
authors = ["Christian <christiandhubbs@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
python-decouple = "^3.7"
whitenoise = "^6.3.0"
django-crispy-forms = "^2.0"
mailchimp-marketing = "^3.0.80"
django-allauth = "^0.52.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
